feat: implement VirtualTable component with resizable columns and expandable rows
- Added VirtualTable component with support for dynamic columns and virtual scrolling.
- Implemented ColumnHeader and ResizableTitle components for customizable column headers.
- Introduced hooks for managing screen width and expanded row states.
- Enhanced data handling with utility functions for SQL keyword extraction and sorting.
- Integrated Ant Design Table with custom expandable rows and sorting capabilities.
- Added types for better TypeScript support and improved code maintainability.
