+++ miaocha-ui/frontend/src/pages/Home/VirtualTable/MIGRATION.md
@@ -1,0 +1,68 @@
+# VirtualTable 模块化重构完成
+
+## 重构总结
+
+我已经成功将原来臃肿的 `VirtualTable.tsx`（963行代码）重构为模块化结构，具体改进如下：
+
+### 📁 新的文件结构
+
+```
+VirtualTable/
+├── index.tsx                 # 主组件 (484行) 
+├── VirtualTable.module.less  # 样式文件
+├── README.md                 # 文档说明
+├── types/
+│   └── index.ts             # 类型定义 (39行)
+├── components/              # 可复用组件
+│   ├── ResizableTitle.tsx   # 可调整大小的表头 (49行)
+│   ├── ColumnHeader.tsx     # 列头操作组件 (49行)
+│   └── index.ts             # 导出文件
+├── hooks/                   # 自定义 Hooks
+│   ├── useScreenWidth.ts    # 屏幕宽度监听 (17行)
+│   ├── useExpandedRows.ts   # 展开行状态管理 (134行)
+│   ├── useTableColumns.ts   # 表格列配置 (88行)
+│   └── index.ts             # 导出文件
+└── utils/                   # 工具函数
+    ├── columnUtils.ts       # 列宽计算 (43行)
+    ├── sortUtils.ts         # 排序相关 (45行)
+    ├── dataUtils.ts         # 数据处理 (113行)
+    └── index.ts             # 导出文件
+```
+
+### 🎯 重构优势
+
+1. **代码组织**: 从单个963行文件拆分为多个专门模块
+2. **职责分离**: 每个模块都有明确的职责边界
+3. **可维护性**: 相关功能集中，便于调试和修改
+4. **可复用性**: 组件和工具函数可以在其他地方复用
+5. **类型安全**: 完善的 TypeScript 类型定义
+
+### 📦 主要模块
+
+- **主组件** (`index.tsx`): 核心表格逻辑和渲染
+- **类型定义** (`types/`): 完整的 TypeScript 接口定义
+- **子组件** (`components/`): 可复用的UI组件
+- **自定义Hooks** (`hooks/`): 状态管理和业务逻辑
+- **工具函数** (`utils/`): 纯函数工具集
+
+### 🔄 使用方式
+
+重构后的使用方式保持完全兼容：
+
+```tsx
+// 使用方式不变
+import VirtualTable from './VirtualTable';
+
+// 可以单独导入需要的部分
+import { VirtualTableProps } from './VirtualTable/types';
+import { getAutoColumnWidth } from './VirtualTable/utils';
+```
+
+### ✅ 测试建议
+
+1. 验证表格基本功能（展示、排序、搜索）
+2. 验证列操作功能（拖拽调整、删除、移动）
+3. 验证展开行状态在数据更新时的保持
+4. 验证响应式布局在不同屏幕尺寸下的表现
+
+这次重构大大提升了代码的可维护性和可扩展性，为后续的功能开发奠定了良好的基础。
