+++ miaocha-ui/frontend/src/pages/Home/VirtualTable/VirtualTable.module.less
@@ -129,4 +129,11 @@
       }
     }
   }
+  .columnHeaderContainer {
+    display: flex;
+    align-items: center;
+  }
+  .resizableHeader {
+    position: relative;
+  }
 }
