+++ miaocha-ui/frontend/src/pages/Home/VirtualTable/components/index.ts
@@ -1,0 +1,3 @@
+// 导出所有组件
+export { default as ResizableTitle } from './ResizableTitle';
+export { default as ColumnHeader } from './ColumnHeader';
