+++ miaocha-ui/frontend/src/pages/Home/VirtualTable/hooks/useScreenWidth.ts
@@ -1,0 +1,19 @@
+import { useState, useEffect } from 'react';
+
+/**
+ * 管理屏幕宽度的Hook
+ */
+export const useScreenWidth = () => {
+  const [screenWidth, setScreenWidth] = useState(window.innerWidth);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setScreenWidth(window.innerWidth);
+    };
+
+    window.addEventListener('resize', handleResize);
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+
+  return screenWidth;
+};
